# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CountsSummary": {
        "additionalProperties": false,
        "description": "Real per-run counts: events, spans, messages, LLM calls and tokens.",
        "properties": {
          "events": {
            "title": "Events",
            "type": "integer"
          },
          "llm_calls": {
            "title": "Llm Calls",
            "type": "integer"
          },
          "messages": {
            "title": "Messages",
            "type": "integer"
          },
          "spans": {
            "title": "Spans",
            "type": "integer"
          },
          "tokens": {
            "title": "Tokens",
            "type": "integer"
          }
        },
        "required": [
          "events",
          "spans",
          "messages",
          "llm_calls",
          "tokens"
        ],
        "title": "CountsSummary",
        "type": "object"
      },
      "DeterminismSummary": {
        "additionalProperties": false,
        "description": "A run's determinism health: canonical log hash and warning counts.",
        "properties": {
          "canonical_log_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Canonical Log Hash"
          },
          "nondeterminism_warnings": {
            "title": "Nondeterminism Warnings",
            "type": "integer"
          },
          "unwrapped_tools": {
            "title": "Unwrapped Tools",
            "type": "integer"
          }
        },
        "required": [
          "canonical_log_hash",
          "unwrapped_tools",
          "nondeterminism_warnings"
        ],
        "title": "DeterminismSummary",
        "type": "object"
      },
...
      "FindingsResponse": {
        "additionalProperties": false,
        "description": "`GET /api/runs/{id}/findings` response \u2014 one run's findings, filtered.",
        "properties": {
          "findings": {
            "items": {
              "$ref": "#/components/schemas/FindingOut"
            },
            "title": "Findings",
            "type": "array"
          }
        },
        "required": [
          "findings"
        ],
        "title": "FindingsResponse",
        "type": "object"
      },
...
      "GraphRef": {
        "additionalProperties": false,
        "description": "A run's graph identity \u2014 content hash plus the observed agent roster.",
        "properties": {
          "agents": {
            "items": {
              "type": "string"
            },
            "title": "Agents",
            "type": "array"
          },
          "hash": {
            "title": "Hash",
            "type": "string"
          }
        },
        "required": [
          "hash",
          "agents"
        ],
        "title": "GraphRef",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "RunDetail": {
        "additionalProperties": false,
        "description": "`GET /api/runs/{id}` response (PRD \u00a726.1, `[SOURCE]`).\n\n`verdict` is `None` until the run's scorecard has been computed and persisted \u2014 no\nprompt in this build yet runs the full `timing -> overhead -> redundancy -> race ->\nbaseline -> resilience -> verdict` pipeline and calls `store.upsert_scorecard` (P17,\nnot started; see `docs/api.md` \"What this endpoint does not yet do\"). This is reported\nas `null`, not fabricated and not a 404 \u2014 PRD \u00a736 rule 2, \"partial results beat no\nresults\": the run's own metadata is always real, even when its verdict is not yet in.",
        "properties": {
          "baseline_run_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline Run Id"
          },
          "comparability": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comparability"
          },
          "counts": {
            "$ref": "#/components/schemas/CountsSummary"
          },
          "determinism": {
            "$ref": "#/components/schemas/DeterminismSummary"
          },
          "graph": {
            "$ref": "#/components/schemas/GraphRef"
          },
          "mode": {
            "title": "Mode",
            "type": "string"
          },
          "run_id": {
            "title": "Run Id",
            "type": "string"
          },
          "scenario": {
            "$ref": "#/components/schemas/ScenarioRef"
          },
          "seed": {
            "title": "Seed",
            "type": "integer"
          },
          "status": {
            "title": "Status",
            "type": "string"
          },
          "timing": {
            "$ref": "#/components/schemas/TimingSummary"
          },
          "verdict": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/VerdictOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "run_id",
          "status",
          "mode",
          "seed",
          "scenario",
          "graph",
          "timing",
          "counts",
          "determinism",
          "verdict",
          "baseline_run_id",
          "comparability"
        ],
        "title": "RunDetail",
        "type": "object"
      },
...
      "ScenarioDetail": {
        "additionalProperties": false,
        "description": "`GET /api/scenarios/{id}` response: raw content plus its resolved form.",
        "properties": {
          "content": {
            "title": "Content",
            "type": "string"
          },
          "content_hash": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Content Hash"
          },
          "path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Path"
          },
          "resolved": {
            "additionalProperties": {
              "$ref": "#/components/schemas/JsonValue-Output"
            },
            "title": "Resolved",
            "type": "object"
          },
          "scenario_id": {
            "title": "Scenario Id",
            "type": "string"
          },
          "source": {
            "enum": [
              "store",
              "file"
            ],
            "title": "Source",
            "type": "string"
          }
        },
        "required": [
          "scenario_id",
          "source",
          "path",
          "content",
          "content_hash",
          "resolved"
        ],
        "title": "ScenarioDetail",
        "type": "object"
      },
...
      "ScenarioRef": {
        "additionalProperties": false,
        "description": "Which scenario a run used \u2014 id (if known) plus its resolved content hash.",
        "properties": {
          "hash": {
            "title": "Hash",
            "type": "string"
          },
          "id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Id"
          }
        },
        "required": [
          "id",
          "hash"
        ],
        "title": "ScenarioRef",
        "type": "object"
      },
...
      "ScenarioValidateRequest": {
        "additionalProperties": false,
        "description": "`POST /api/scenarios/validate` request body.\n\nExactly one of `text` (raw YAML, e.g. from the UI's chaos panel) or `scenario_id` (an\nalready-known scenario, from the store or `scenarios_dir`) must be given.",
        "properties": {
          "scenario_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Scenario Id"
          },
          "source_name": {
            "default": "<scenario>",
            "title": "Source Name",
            "type": "string"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          }
        },
        "title": "ScenarioValidateRequest",
        "type": "object"
      },
      "ScenarioValidateResponse": {
        "additionalProperties": false,
        "description": "`POST /api/scenarios/validate` response \u2014 always `200`, see the route's docstring.",
        "properties": {
          "errors": {
            "items": {
              "$ref": "#/components/schemas/ScenarioValidationErrorOut"
            },
            "title": "Errors",
            "type": "array"
          },
          "resolved": {
            "anyOf": [
              {
                "additionalProperties": {
                  "$ref": "#/components/schemas/JsonValue-Output"
                },
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Resolved"
          },
          "valid": {
            "title": "Valid",
            "type": "boolean"
          },
          "warnings": {
            "default": [],
            "items": {
              "type": "string"
            },
            "title": "Warnings",
            "type": "array"
          }
        },
        "required": [
          "valid",
          "resolved",
          "errors"
        ],
        "title": "ScenarioValidateResponse",
        "type": "object"
      },
      "ScenarioValidationErrorOut": {
        "additionalProperties": false,
        "description": "One `scenario.validate.ScenarioError`, projected.",
        "properties": {
          "code": {
            "title": "Code",
            "type": "string"
          },
          "docs": {
            "title": "Docs",
            "type": "string"
          },
          "file": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File"
          },
          "line": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Line"
          },
          "message": {
            "title": "Message",
            "type": "string"
          },
          "path": {
            "title": "Path",
            "type": "string"
          },
          "suggestion": {
            "title": "Suggestion",
            "type": "string"
          }
        },
        "required": [
          "code",
          "path",
          "message",
          "suggestion",
          "file",
          "line",
          "docs"
        ],
        "title": "ScenarioValidationErrorOut",
        "type": "object"
      },
...
      "TimingSummary": {
        "additionalProperties": false,
        "description": "Makespan figures for `GET /runs/{id}` \u2014 virtual, wall and critical-path ms.",
        "properties": {
          "critical_path_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Critical Path Ms"
          },
          "virtual_makespan_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Virtual Makespan Ms"
          },
          "wall_makespan_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Wall Makespan Ms"
          }
        },
        "required": [
          "virtual_makespan_ms",
          "wall_makespan_ms",
          "critical_path_ms"
        ],
        "title": "TimingSummary",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
      "VerdictOut": {
        "additionalProperties": false,
        "description": "`analysis.verdict.Verdict`, projected \u2014 `GET /runs/{id}`'s `verdict` object (PRD \u00a718).",
        "properties": {
          "class": {
            "title": "Class",
            "type": "string"
          },
          "confidence": {
            "title": "Confidence",
            "type": "string"
          },
          "coordination_score": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Coordination Score"
          },
          "evidence": {
            "$ref": "#/components/schemas/Evidence"
          },
          "findings": {
            "items": {
              "$ref": "#/components/schemas/VerdictFinding"
            },
            "title": "Findings",
            "type": "array"
          },
          "headline": {
            "title": "Headline",
            "type": "string"
          },
          "recommendations": {
            "items": {
              "$ref": "#/components/schemas/Recommendation"
            },
            "title": "Recommendations",
            "type": "array"
          },
          "secondary_classes": {
            "items": {
              "type": "string"
            },
            "title": "Secondary Classes",
            "type": "array"
          }
        },
        "required": [
          "class",
          "secondary_classes",
          "coordination_score",
          "confidence",
          "headline",
          "findings",
          "recommendations",
          "evidence"
        ],
        "title": "VerdictOut",
        "type": "object"
      },
...
    }
  },
  "info": {
    "summary": "Multi-agent coordination debugger, deterministic replay runtime and chaos harness \u2014 PRD \u00a726 API layer. Bounded search: absence of findings is not proof of absence.",
    "title": "AgentDX API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/runs/{run_id}": {
      "get": {
        "description": "Return a run's metadata, real counts and (when computed) its verdict.\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.\n\n`timing.critical_path_ms` is reported as `null`: computing it means building the full\ntiming DAG (`analysis.timing.build_timing_dag` + `critical_path`), the same cost\n`GET .../graph` already pays where the PRD's own example payload needs the number \u2014 this\nendpoint does not additionally pay that cost on every call just to fill one field callers\nneeding it can get from `.../graph` directly. `determinism.unwrapped_tools` is `0` on\nevery run: no event or field anywhere in this codebase currently records it (grepped, not\nassumed) \u2014 a real, declared gap, not a fabricated zero passed off as a real count.",
        "operationId": "get_run_detail_api_runs__run_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RunDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run metadata + verdict (PRD \u00a726.1)",
        "tags": [
          "runs"
        ]
      }
    },
...
    "/api/runs/{run_id}/findings": {
      "get": {
        "description": "Return a run's findings, severity/type-filtered, precomputed (PRD \u00a726.1, \u00a726.3).\n\nRaises:\n    RunNotFoundError: `E-RUN-404`.",
        "operationId": "get_findings_api_runs__run_id__findings_get",
        "parameters": [
          {
            "in": "path",
            "name": "run_id",
            "required": true,
            "schema": {
              "title": "Run Id",
              "type": "string"
            }
          },
          {
            "description": "Filter: exact severity match.",
            "in": "query",
            "name": "severity",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter: exact severity match.",
              "title": "Severity"
            }
          },
          {
            "description": "Filter: exact type match.",
            "in": "query",
            "name": "type",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter: exact type match.",
              "title": "Type"
            }
          },
          {
            "description": "Include findings a guard suppressed.",
            "in": "query",
            "name": "include_suppressed",
            "required": false,
            "schema": {
              "default": false,
              "description": "Include findings a guard suppressed.",
              "title": "Include Suppressed",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingsResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Findings (PRD \u00a726.1)",
        "tags": [
          "findings"
        ]
      }
    },
...
    "/api/scenarios/validate": {
      "post": {
        "description": "Validate scenario text or an existing scenario id and return the resolved document.\n\nNever raises for an invalid scenario (Design Constraint 1's \"thin\" still holds: this\ncalls `scenario.loader`/`scenario.validate` and serialises, no rule lives here) \u2014 a\nmalformed or invalid scenario is a `200` with `valid: false`, because that is what makes\nthis endpoint safe for the UI's chaos panel to call on every keystroke.\n\nRaises:\n    ScenarioNotFoundError: `scenario_id` was given and names nothing in either source.",
        "operationId": "validate_scenario_api_scenarios_validate_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScenarioValidateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioValidateResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Dry-run validate a scenario (PRD Sec26.1)",
        "tags": [
          "scenarios"
        ]
      }
    },
    "/api/scenarios/{scenario_id}": {
      "get": {
        "description": "Return one scenario's raw content plus its resolved (defaults-applied) form.\n\nRaises:\n    ScenarioNotFoundError: `E-SCEN-404` \u2014 neither source has this id.",
        "operationId": "get_scenario_api_scenarios__scenario_id__get",
        "parameters": [
          {
            "in": "path",
            "name": "scenario_id",
            "required": true,
            "schema": {
              "title": "Scenario Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScenarioDetail"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Fetch a scenario (PRD Sec26.1)",
        "tags": [
          "scenarios"
        ]
      }
    },
...
  }
}